+++ CHANGELOG.md
@@ -9,3 +9,8 @@
 
 - `npm install --save siwe ethers`
 - Added types/index.ts to give typing to `window.ethereum`
+
+## Step 3
+
+- creating nonce through web browser (no longer using the sample code from siwe)
+- disabled button for verify